feature: 차단 설정 화면 스크롤 페이징, 목록 없을 때 처리

+++ CatchMate/data/src/main/java/com/catchmate/data/dto/user/GetBlockedUserListResponseDTO.kt
@@ -3,7 +3,7 @@ package com.catchmate.data.dto.user
 data class GetBlockedUserListResponseDTO(
     val userInfoList: List<GetUserProfileResponseDTO>,
     val totalPages: Int,
-    val totalElements: Long,
+    val totalElements: Int,
     val isFirst: Boolean,
     val isLast: Boolean,
 )
